Fail travis build if codegen isn't committed
This doesn't catch 100% of such problems, as it runs before tests have
been run. However, almost all of our protos are used in 'main' so this
limitation shouldn't be a big deal.

If it does become a problem, we change it from 'before_script' to
'script', but then we'll also need to include gradle build instructions.